+++ tests/test_numbers.py
@@ -49,7 +49,41 @@ class TestNumbers(unittest.TestCase):
                 random_number = choice(org_number_list)
             self.assertIsInstance(random_number, Number)
             got_number = org_number_list.get(extension=random_number.extension)
-            self.assertEqual(random_number, got_number)
+            self.assertIsInstance(got_number, (Number, list))
+            if isinstance(got_number, Number):
+                self.assertEqual(random_number, got_number)
+        # The following looks for an extension that exists at more than one Location, picks the Location and searches
+        # for the extension, confirming it got the right Location
+        with self.subTest('Get Number by Extension and Location'):
+            seen_extensions = []
+            selected_extension = None
+            selected_location = None
+            for number in org_number_list:
+                if number.extension is not None:
+                    if number.extension in seen_extensions:
+                        selected_extension = number.extension
+                        selected_location = number.location
+                        break
+                    else:
+                        seen_extensions.append(number.extension)
+            self.assertIsInstance(selected_extension, str)
+            self.assertIsInstance(selected_location, wxcadm.Location)
+            found = self.webex.org.numbers.get(extension=selected_extension, location=selected_location)
+            self.assertIsInstance(found, Number)
+            self.assertEqual(selected_extension, found.extension)
+            self.assertEqual(selected_location, found.location)
+        with self.subTest('Get Number by Owner'):
+            random_user: Optional[wxcadm.Person] = None
+            org_people = self.webex.org.people.webex_calling()
+            while random_user is None or random_user.numbers is None:
+                random_user = choice(org_people)
+            self.assertIsInstance(random_user, wxcadm.Person)
+            number = self.webex.org.numbers.get_by_owner(random_user)
+            self.assertIsInstance(number, wxcadm.Number)
+            self.assertEqual(random_user, number.owner)
+
+
+
 
 
 if __name__ == '__main__':